% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocyc-transform.R
\name{biocyc_transform}
\alias{biocyc_transform}
\title{Convert BioCyc SmartTable}
\usage{
biocyc_transform(session, table_id, transform_id, index = 0)
}
\arguments{
\item{session}{httr::handle object, representing the established BioCyc session}

\item{table_id}{string, unique identifier of SmartTable}

\item{transform_id}{string, the ID of the transformation operation to be performed}

\item{index}{integer, specifying the index of the column to be converted, the default is 0}
}
\value{
No return value, a success message will be printed when the function is executed successfully
}
\description{
This function is used to perform specified conversion operations on BioCyc SmartTable.
}
\details{
There are a total of 104 transformation operations,
which can be obtained through \code{cailab.utils:::biocyc_get_transformations()}. Users can view the details of them
in the \code{transformations} data by \code{data("transformations")}.
}
\examples{
\dontrun{
session <-biocyc_session("your_email@example.com", "your_password")
biocyc_transform(session, "your_table_id", "your_transform_id")
}

}
\seealso{
For details on available transformation operations, see https://biocyc.org/st-transformations-table
}
