% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliometrix.R
\name{DT_output}
\alias{DT_output}
\title{Use \code{DT} to output a tibble}
\usage{
DT_output(
  M,
  caption = "",
  filename = knitr::fig_path(),
  column = c("TI", "DI", "TC")
)
}
\arguments{
\item{M}{a data.frame}

\item{caption}{table caption}

\item{filename}{download file of table}

\item{column}{columns to show in M}
}
\description{
Use \code{DT} to output a tibble
}
