% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown-utils.R
\name{figure_caption_link}
\alias{figure_caption_link}
\title{Show the download links of figures when multiple devices are used}
\usage{
figure_caption_link(..., PPT = TRUE, PDF = TRUE)
}
\arguments{
\item{...}{figure caption}

\item{PPT}{whether output PPT link}

\item{PDF}{whether output PDF link}
}
\value{
a markdown format string
}
\description{
When use multiple devices, i.e., pdf, svg, ppt, multiple figures will
be generated after rendering. This function automatically add the links
to these files in figure captions.
}
