% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_citation.R
\name{get_pubmed_record}
\alias{get_pubmed_record}
\title{从 PubMed 获取文献记录}
\usage{
get_pubmed_record(pubmed_id)
}
\arguments{
\item{pubmed_id}{字符串，PubMed ID。}
}
\value{
列表，包含文献的详细信息。
}
\description{
该函数使用 PubMed ID 从 PubMed 数据库获取文献信息。
}
\examples{
\dontrun{
record = get_pubmed_record("12345678")
print(record)
}

}
