% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-plot.R
\name{gg_boxplot_with_group}
\alias{gg_boxplot_with_group}
\title{Plot a boxplot and add group to it}
\usage{
gg_boxplot_with_group(
  data,
  x,
  y,
  method = "aov",
  with_jitter = TRUE,
  label_repel = TRUE
)
}
\arguments{
\item{data}{data}

\item{x}{group variable, should be a column in data}

\item{y}{values, should be a column in data}

\item{method}{aov}

\item{with_jitter}{add jitter points}

\item{label_repel}{use ggrepel to plot label}
}
\value{
a ggplot object
}
\description{
See also: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6193594/
}
\examples{
 data("PlantGrowth")
 gg_boxplot_with_group(PlantGrowth, "group", "weight")
}
