% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliometrix-network.R
\name{graph_add_node}
\alias{graph_add_node}
\alias{graph_add_node_pagerank}
\alias{graph_add_node_degree}
\title{Modification of igraph Object}
\usage{
graph_add_node_pagerank(g)

graph_add_node_degree(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
a new igraph object
}
\description{
Modification of igraph Object
}
