% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown-utils.R
\name{kable_classic2}
\alias{kable_classic2}
\title{A table generated by kableExtra}
\usage{
kable_classic2(tibble)
}
\arguments{
\item{tibble}{a tibble}
}
\value{
a table
}
\description{
Can be safely used in PDF document.
}
\examples{
kable_classic2(head(iris))
}
