% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence-alignment.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot tree with ggtree}
\usage{
plot_tree(tree, tip.label = "label", tip.color = "parent")
}
\arguments{
\item{tree}{tree}

\item{tip.label}{tip.label mapping}

\item{tip.color}{tip.color mapping}
}
\value{
a ggplot object
}
\description{
By default, this function plot a tree with tip labels and colors.
}
