% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocyc-data.R
\name{properties}
\alias{properties}
\title{BioCyc 数据库属性说明}
\format{
一个 tibble 数据框,包含以下列:
\describe{
\item{id}{属性的唯一标识符}
\item{description}{属性的详细说明}
}
}
\source{
BioCyc 数据库文档
}
\description{
BioCyc 数据库中可用的属性列表及其说明
}
