% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-proxy.R
\name{set_proxy}
\alias{set_proxy}
\title{Set http(s) proxy in the current R session}
\usage{
set_proxy(url = "http://127.0.0.1", port = 36119)
}
\arguments{
\item{url}{proxy IP address}

\item{port}{proxy port}
}
\value{
silent
}
\description{
This function is useful when you network is not able to access needed content by default.
For instance, when you are install package from GitHub, but you are blocked by firewalls.
}
\details{
Try to find the port automatically.
}
\examples{
  set_proxy()
  # run command
  # tinytex::install()
}
