% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliometrix-network.R
\name{simplified_network}
\alias{simplified_network}
\alias{country_network}
\alias{author_network}
\alias{university_network}
\alias{keyword_network}
\title{Construct Networks of Different Tags}
\usage{
simplified_network(
  M,
  from = NULL,
  to = NULL,
  nNode = 30,
  remove_keyword = NULL,
  edge_weight_cutoff = 1,
  analysis,
  network,
  field,
  delete_isolate = TRUE,
  graph = FALSE,
  ...
)

country_network(
  M,
  analysis = "collaboration",
  network = "countries",
  field = "AU_CO_NR",
  edge_weight_cutoff = 5,
  nNode = 20,
  graph = FALSE,
  ...
)

author_network(
  M,
  analysis = "collaboration",
  network = "authors",
  field = "AU",
  edge_weight_cutoff = 5,
  nNode = 200,
  graph = FALSE,
  ...
)

university_network(
  M,
  analysis = "collaboration",
  network = "universities",
  field = "AU_UN_NR",
  edge_weight_cutoff = 10,
  nNode = 30,
  graph = FALSE,
  ...
)

keyword_network(
  M,
  nNode = 100,
  edge_weight_cutoff = 3,
  field = "ID",
  analysis = "co-occurrences",
  network = "keywords",
  graph = FALSE,
  ...
)
}
\arguments{
\item{M}{bibliometrix data frame}

\item{from}{start of PY}

\item{to}{end of PY}

\item{nNode}{Maximum number of nodes presented in the network}

\item{remove_keyword}{regex used to filter data}

\item{edge_weight_cutoff}{edge weight cutoff}

\item{analysis}{type of analysis}

\item{network}{type of network}

\item{field}{data column for network construction}

\item{delete_isolate}{TRUE}

\item{graph}{whether return graph}

\item{...}{pass to \code{biblio_network()}}
}
\value{
visNetwork object
}
\description{
Construct Networks of Different Tags
}
\examples{
\dontrun{
library("bibliometrixData")
data("garfield")
author_network(garfield)
}
}
