% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliometrix-plot.R
\name{tableTag_barplot}
\alias{tableTag_barplot}
\title{tableTag downstream plot}
\usage{
tableTag_barplot(M, Tag = "AU", n = 30)
}
\arguments{
\item{M}{a bibliometrix data.frame}

\item{Tag}{passed to \code{tableTag()}}

\item{n}{top n}
}
\value{
a ggplot object
}
\description{
tableTag downstream plot
}
\examples{
library("bibliometrixData")
data("garfield")
tableTag_barplot(garfield, Tag = "AU")
}
