% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence-alignment.R
\name{tree_from_aln}
\alias{tree_from_aln}
\title{Construct a phylogenetic tree from sequence alignment}
\usage{
tree_from_aln(aln, dist.method = "hamming")
}
\arguments{
\item{aln}{a 'DNAMultipleAlignment' object}

\item{dist.method}{method to caculate distance}
}
\value{
a tree
}
\description{
Construct a phylogenetic tree from sequence alignment
}
