% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown-utils.R
\name{var2header}
\alias{var2header}
\title{Remove specific connector in string}
\usage{
var2header(x)
}
\arguments{
\item{x}{string}
}
\value{
a string
}
\description{
变量名中间常用下划线连接单词，而这在显示的时候是多余。
这个函数可以去掉单词间的下划线。
}
