% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_summary.R
\name{yt_playlist_videos}
\alias{yt_playlist_videos}
\title{获取播放列表中的视频 URL}
\usage{
yt_playlist_videos(playlist_url)
}
\arguments{
\item{playlist_url}{字符串,播放列表 URL}
}
\value{
字符向量,包含所有视频的 URL
}
\description{
获取播放列表中的视频 URL
}
\keyword{internal}
