% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhipuai-batch-mode.R
\name{zhipuai_batch_list_all}
\alias{zhipuai_batch_list_all}
\title{迭代获取所有 Batch 任务}
\usage{
zhipuai_batch_list_all(limit = 20, api_key = Sys.getenv("ZHIPUAI_API_KEY"))
}
\arguments{
\item{limit}{每页返回的数量,默认20}

\item{api_key}{智谱AI的API密钥,默认从环境变量ZHIPUAI_API_KEY中获取}
}
\value{
返回一个包含所有批处理任务的数据框
}
\description{
此函数用于迭代获取智谱AI平台上的所有批处理任务
}
