% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhipuai-result-parser.R
\name{zhipuai_result_chat_parser}
\alias{zhipuai_result_chat_parser}
\title{解析智谱 AI 任务结果中的聊天任务结果}
\usage{
zhipuai_result_chat_parser(result)
}
\arguments{
\item{result}{智谱 AI 聊天模型的任务结果}
}
\value{
返回一个数据框
}
\description{
解析智谱 AI 任务结果中的聊天任务结果
}
